% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extend_seq_range.R
\name{extend_seq_range}
\alias{extend_seq_range}
\title{Extend assay sequences using a reference genome}
\usage{
extend_seq_range(tbl, len)
}
\arguments{
\item{tbl}{A table of assays with columns: locus, acc, gstart, gend, reverse, pos, ref, alt, orig_seq}

\item{len}{Number of bases to extend on either side of the original sequence}
}
\value{
A tibble with the extended sequence
}
\description{
Extend assay sequences using a reference genome
}
