% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_seq.R
\name{fetch_seq}
\alias{fetch_seq}
\title{Fetch a nucleotide sequence from NCBI or a FASTA file}
\usage{
fetch_seq(accession = NULL, fasta = NULL, seq_name = NULL, seq_start, seq_end)
}
\arguments{
\item{accession}{Sequence accession number from NCBI nuccore}

\item{fasta}{Path to the fasta file}

\item{seq_name}{The name of the sequence in the FASTA file}

\item{seq_start}{Start coordinate}

\item{seq_end}{End coordinate}
}
\value{
The requested sequence (character string)
}
\description{
Fetch a nucleotide sequence from NCBI or a FASTA file
}
