% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_snp_targets.R
\name{find_snp_targets}
\alias{find_snp_targets}
\title{Find short microhaplotypes for primer design}
\usage{
find_snp_targets(df, span = 60)
}
\arguments{
\item{df}{A data frame with columns 'seq_name' and 'pos'}

\item{span}{The maximum distance between SNPs when choosing microhaplotypes}
}
\value{
A tibble with columns 'seq_name' and 'target'
}
\description{
Find short microhaplotypes for primer design
}
