% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_snp_position.R
\name{get_snp_position}
\alias{get_snp_position}
\title{Locate the position of a SNP in a sequence with assay notation}
\usage{
get_snp_position(seq)
}
\arguments{
\item{seq}{A vector of sequences in assay notation '\\[C/T\\]'}
}
\value{
An integer representing the location of the SNP in the provided sequences
}
\description{
Locate the position of a SNP in a sequence with assay notation
}
