% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_blast.R
\name{read_blast}
\alias{read_blast}
\title{Import a BLAST report as a tibble}
\usage{
read_blast(file, format = "tab")
}
\arguments{
\item{file}{Path to the BLAST report file}

\item{format}{Currently only supports tab separated reports (outfmt = "6")}
}
\value{
A tibble with BLAST results
}
\description{
Import a BLAST report as a tibble
}
\note{
Columns:
1.    qseqid 	 query (e.g., gene) sequence id
2. 	 sseqid 	 subject (e.g., reference genome) sequence id
3. 	 pident 	 percentage of identical matches
4. 	 length 	 alignment length
5. 	 mismatch 	 number of mismatches
6. 	 gapopen 	 number of gap openings
7. 	 qstart 	 start of alignment in query
8. 	 qend 	 end of alignment in query
9. 	 sstart 	 start of alignment in subject
10. 	 send 	 end of alignment in subject
11. 	 evalue 	 expect value
12. 	 bitscore  bit score
}
