% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fasta.R
\name{read_fasta}
\alias{read_fasta}
\title{Import a FASTA file into a tibble}
\usage{
read_fasta(fasta)
}
\arguments{
\item{fasta}{Path to the input FASTA file}
}
\value{
A tibble with columns 'seq_id' and 'seq'
}
\description{
Import a FASTA file into a tibble
}
