% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_assay_notation.R
\name{remove_assay_notation}
\alias{remove_assay_notation}
\title{Convert DNA sequences with assay notation to regular sequences}
\usage{
remove_assay_notation(seq)
}
\arguments{
\item{seq}{DNA sequences in assay notation (i.e. "CCTG\\[T/G\\]TAGG")}
}
\value{
A vector of sequences without assay notation
}
\description{
Convert DNA sequences with assay notation to regular sequences
}
