% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/self_blast.R
\name{self_blast}
\alias{self_blast}
\title{BLAST-compare a set of sequences to identify regions of similarity}
\usage{
self_blast(seq, id, conda_env = CONDA_ENV, conda_sh_path = CONDA_PATH)
}
\arguments{
\item{seq}{A vector of sequences}

\item{id}{A vector of sequence names}

\item{conda_env}{The conda environment}

\item{conda_sh_path}{The path to the conda.sh}
}
\value{
A tibble with regions of similarity
}
\description{
BLAST-compare a set of sequences to identify regions of similarity
}
