% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqhelp-package.R
\docType{package}
\name{seqhelp-package}
\alias{seqhelp}
\alias{seqhelp-package}
\title{seqhelp: Utilities for DNA Sequence Manipulation}
\description{
Provides miscellaneous functions for manipulating DNA sequences using tidy principles
}
\author{
\strong{Maintainer}: Chris Hollenbeck \email{chollenbeck07@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
