% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/systemc.R
\name{systemc}
\alias{systemc}
\title{Run a system call inside a conda environment}
\usage{
systemc(
  command,
  env = CONDA_ENV,
  ignore_stdout = FALSE,
  ignore_stderr = FALSE,
  conda_sh_path = CONDA_PATH
)
}
\arguments{
\item{command}{The system command (Linux only)}

\item{env}{The conda environment}

\item{ignore_stdout}{Ignore STDOUT (TRUE/FALSE)}

\item{ignore_stderr}{Ignore STDERR (TRUE/FALSE)}

\item{conda_sh_path}{The path to the conda.sh}
}
\value{
The return value from the system call
}
\description{
Run a system call inside a conda environment
}
