% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_fasta.R
\name{write_fasta}
\alias{write_fasta}
\title{Write a FASTA file}
\usage{
write_fasta(file, names, seq)
}
\arguments{
\item{file}{The path of the FASTA file to create}

\item{names}{Vector of sequence names (to become the FASTA ID)}

\item{seq}{Vector of sequences}
}
\value{
Nothing
}
\description{
Write a FASTA file
}
