% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_primer3.R
\name{run_primer3}
\alias{run_primer3}
\title{Design primers for a sequence}
\usage{
run_primer3(
  df,
  amplicon_length = c(110, 150, 190),
  melting_temp = c(60, 60, 65),
  primer_length = c(20, 20, 25),
  settings_file = NULL,
  settings = NULL
)
}
\arguments{
\item{df}{A dataframe with at least the following columns: seq, seq_name, target_string
seq: A character string containing the sequence
seq_name: The name of the sequence
target_string: String to identify the target region - takes the form "target,length"}

\item{amplicon_length}{Vector of length three with (min, opt, max) of amplicon product size}

\item{melting_temp}{Vector of length three with (min, opt, max) of melting temperature (Tm)}

\item{primer_length}{Vector of length three with (min, opt, max) of primer length}

\item{settings_file}{Optional path to a file of predetermined settings}

\item{settings}{Optional named list of additional settings}
}
\value{

}
\description{
Design primers for a sequence
}
